1.Common.BaseEntity，BaseVO改为接口形式。添加BaseEntityWithId实现BaseEntity，描述主键为id的表。添加BaseVOImpl实现BaseVO，BaseVOWithIdImpl继承BaseVOImpl，描述主键为id的表对应的vo类。 2.添加订单模块、商品模块 3.完成 下单->减库存->生成订单 的业务逻辑，通过了分布式事务的测试。

+++ project/Commons/src/main/java/com/zhoushiya/shop/common/dao/BaseDao.java
@@ -13,7 +13,7 @@ public interface BaseDao<Entity extends BaseEntity> {
 
     int create(Entity record);
 
-    Entity selectByPrimaryKey(Integer id);
+    Entity selectByPrimaryKey(long id);
 
     int updateByPrimaryKey(Entity record);
